Fix highlighting of counters
The change introduced in 794e229dde2f02a1288eef6c8395b34870714985
doesn't work because it iterates over the wrong elements on the profile
page.
This commit iterates directly over all 'count-number' elements and adds
a new class for highlighting if the element is not equal to '0'. This
means that it does not only fix the bug but also simplifies the code a
lot.